+++ Forge/src/main/java/net/potionstudios/biomeswevegone/forge/VanillaCompatForge.java
@@ -15,6 +15,7 @@ import net.minecraftforge.event.entity.player.BonemealEvent;
 import net.minecraftforge.event.furnace.FurnaceFuelBurnTimeEvent;
 import net.minecraftforge.event.level.BlockEvent;
 import net.minecraftforge.event.village.VillagerTradesEvent;
+import net.minecraftforge.event.village.WandererTradesEvent;
 import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.potionstudios.biomeswevegone.util.BoneMealHandler;
@@ -47,10 +48,18 @@ public class VanillaCompatForge {
     public static void registerVanillaCompatEvents(IEventBus bus) {
         bus.addListener(VanillaCompatForge::registerTillables);
         bus.addListener(VanillaCompatForge::registerFuels);
-        if (BWGTradesConfig.INSTANCE.get().enableTrades()) bus.addListener(VanillaCompatForge::onVillagerTrade);
+        if (!BWGTradesConfig.INSTANCE.trades.disableTrades.value()) {
+            bus.addListener(VanillaCompatForge::onVillagerTrade);
+            if (BWGTradesConfig.INSTANCE.wanderingTraderTrades.enableBWGItemsTrades.value())
+                bus.addListener(VanillaCompatForge::onWanderingTrade);
+        }
         bus.addListener(VanillaCompatForge::onBoneMealUse);
     }
 
+    /**
+     * Register tillable blocks.
+     * @see BlockEvent.BlockToolModificationEvent
+     */
     private static void registerTillables(final BlockEvent.BlockToolModificationEvent event) {
         if (event.getToolAction() == ToolActions.HOE_TILL && event.getLevel().getBlockState(event.getPos().above()).isAir()) {
             BlockState state = event.getState();
@@ -85,6 +94,16 @@ public class VanillaCompatForge {
                             for (MerchantOffer offer : offers) tradeList.add((trader, random) -> offer);
                     });
         }
+    }
+
+    /**
+     * Register wandering trader trades.
+     * @see WandererTradesEvent
+     */
+    private static void onWanderingTrade(final WandererTradesEvent event) {
+        BWGVillagerTrades.WANDERING_TRADER_TRADES.forEach((level, offers) -> {
+            for (MerchantOffer offer : offers) event.getGenericTrades().add((trader, random) -> offer);
+        });
     }
 
     /**